refactor(sessions): extract metadata registry
