+++ AGENTS.md
@@ -81,6 +81,7 @@ Tracked-client identity lives in **one** place: `CLIENT_CATALOG` in `src/shared/
 |---|---|
 | Client identity | one entry in `CLIENT_CATALOG` (`src/shared/clientCatalog.js`): id, label, display position, `defaultTracked`, `locallyParsed`. `DEFAULT_CLIENTS` / `KNOWN_CLIENTS` / `PARSE_LOCAL_CLIENTS` in `clientTracking.js` and the renderer's `clientLabels` / `KNOWN_CLIENTS` are all derived from it, so the tracked-client id, label and display order used by tracking and the widget renderer are declared once — Discord's `CLIENT_LABELS` and `themePresets`'s `VENDOR_LABELS` still carry their own |
 | Client-specific code | only when the client parses or syncs something itself (transcript readers, path resolvers, the self-sync a cache-backed client needs): `src/shared/providers/<id>/`, the same folder the limits provider of that id would use. Most tracked clients need none — tokscale parses them end to end, so they are catalog, roots, icon and README only |
+| Session metadata | only when Session Detail rows need local metadata beyond tokscale's usage record: add one resolver entry in `src/shared/sessionMetadata.js`. Every resolver receives the requested session-id set plus shared context and returns `Map<sessionId, { startedAt, lastUsedAt, projectId, projectLabel, title, sessionKind }>`; keep storage-specific discovery, parsing, cache, and refresh policy in `src/shared/providers/<id>/` rather than inventing one cross-provider cache abstraction |
 | Source roots | the `add(...)` call in `clientSourceRoots()` (`src/shared/collector.js`) — one `[checkId, dir]`, or `[checkId, watchDir, sourcePath]` when tokscale reads one exact file. `clientWatchCandidates()` is only a projection of this table; nothing is declared there |
 | Source check ids | every `checkId` above must be in `CLIENT_SOURCE_CHECK_IDS` (`src/shared/clientHealth.js`), kept alphabetical, then `npm run sync:worker` for the Worker copy. An id missing from that allowlist makes `normalizeClientHealth` drop the client's whole `checks` array, not just the unknown entry |
 | XDG vs home-relative | mirror tokscale, do not guess: a root is XDG-derived only if `clients.rs` declares it `PathRoot::XdgData` or `scanner.rs` resolves it through the `dirs` crate. Those `dirs` lookups are invisible to `strings` on the binary and to `tokscale clients`, so read the Rust at the version tag (`tmp/tokscale`). Roots spelled as home-relative literals upstream must stay home-relative here |
